Various works to create the subject detail fragment.
There's also some design works added, but they're not final.

+++ app/src/main/java/black/old/spacedrepetitionowl/database/ReminderDao.kt
@@ -25,7 +25,7 @@ interface ReminderDao {
 
     // No need for coroutines when LiveData is used.
     @Query("SELECT * FROM sro_reminders WHERE subjectId = :subjectId ORDER BY dateTimestamp ASC")
-    fun getRemindersBySubject(subjectId: Int): LiveData<List<Reminder>>
+    fun getRemindersBySubject(subjectId: Long): LiveData<List<Reminder>>
 
     // Empty out table
     @Query("DELETE FROM sro_reminders")
